+++ CHANGELOG.md
@@ -31,6 +31,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
   packages (171) has to be added on top of it, starting with the smaller image
   is preferable. (See [#0.9.6---2020-08-13](v0.9.6 below) for why not using one
   of the `python` base images.)
+- Default for `--push-branch` is now `*`, meaning `--all` (which cannot be
+  expressed in the config file)
 
 
 # 1.0.2 - 2020-08-15